requirements.txt
PyQt6>=6.4.0
pillow>=10.0.0